COMPAT_43 KERNEL DEPENDENCY CLEANUP
	 Here is a patch for the use of COMPAT_43 in the 386BSD kernel.
This is an extremely LOW priority patch, because lots of applications
(init, csh, vi)	depend on this being defined, so you can't use a kernel
with this option removed. The only application that does cope with this
option missing is my own shell which is POSIX-compliant. But you might
want to add it in now rather than have to rediscover it at a later
stage.

AUTHOR:		Warren Toomey
386BSD-Patchkit: patch00062