Fix yp_mkdb to do what the code suggests it's trying to do..
The code, as written, appears to load the new database data into a new
hash file and renames the two.

Due to a run of bugs and lack of error checking, it's going a whole
mess of unlink() and rename() calls that are failing.  It only
worked in the first place because the data was being inserted into a
"live" hash file.  (I wonder how much stale data has assumulated?)

Submitted by: Laurence Lopez <lopez@mv.mv.com>