When a timeval is stored at the beginning of icmp_data, the fields are unaligned for access by the alpha, so copy the value to a variable that is aligned.
When checking the returned data, be careful to avoid confusing the
size of the icmp header with the size of a timeval. On i386 these
are both 8, but on alpha, a timeval is 16 bytes. This means that
a packet sent from an alpha contains 48 bytes of data, not 56 like
on i386.