It's not been possible to configure a print server running lpd, with remote hosts feeding it, so that some hosts have their header pages supressed and some don't.  This is because lpd doesn't know how to rewrite a print job before forwarding it to a remote lpd. In particular this causes problems with p rinters that contain their own lpd, eg. HP jet direct cards, because they can't suppress headers.  It's not possible to have headers supressed by putting 'sh' in any printcap in the lpd chain, it is up to the originating lpr to have a '-h' option specified at run time.
Lpr has been modified to allow _it_ to honour the 'sh' flag in the
local print cap.  This allows the administrator to switch off
headers for a particular printer (on a particular host) irrespective
of whether that printer is local to the machine or remote.

This doesn't break anything, because in the case of a remote printer
the 'sh' flag would have had no meaning, in the case of the local
printer it would have been on anyway.

Submitted by:	Scott James Remnant <scott@pavilion.net>