MFC: 1.122->1.223 Set up `ia' properly for dummynet packets to avoid dereferencing an unitialised pointer.
This change includes some over-defensive checks not present in the
patch applied to -current; we set `ia' to NULL at the start of
ip_output(), check that the dummynet route is non-NULL, and that
`ia' is non-NULL before it is dereferenced. This are just safety
measures which can be removed when the per-ifaddr stats code has
had a bit more testing time in -current.

Approved by:	jkh