New release notes:  Use net/cvsup, rather than net/cvsup-bin and net/cvsupd-bin.  Also note CVSup update to fix S1G bug.
I'm breaking a self-imposed rule of not mentioning ports-collection
updates in release notes.  I'm only doing this because CVSup is a
fairly essential part of the FreeBSD Project infrastructure.  Save
for exceptions such as this, the release notes will continue to
document changes to the base system only.

Suggested by:	mike