o nvi failed to correctly handle EINTR returned from tcsetattr(), exiting   instead of recovering, which happens in particular during a rapid series   of SIGWINCH's.  This change forces nvi to loop on the call in the event   that the call is interrupted.
Interestingly, I submitted this bug report in 1998, and a solution was
posted shortly thereafter by Matthias Buelow; it's been sitting in the
PR database ever since.

Note: this takes this file off the vendor branch.  If and when we find
a vendor for this code, the fix should be given back to them.

PR:		bin/8438
Submitted by:	Matthias Buelow <mkb@altair.mayn.de>
MFC after:	1 month