Hacks for measuring interrupt latency.  Interrupt latency can be measured accurately for periodic interrupts provided the interrupts don't need to be serviced very quickly to keep their period almost constant.  sio output interrupts have this property (interrupt service can be delayed for up to 1 character time without the period changing).
This is non-optional and undocumented so that it can be added and
removed easily.  It has no significant effect unless it is enabled by
hacking on a variable using a debugger.  Hardclock and statclock interrupts
would work even better for this, at least on i386's, provided their
interrupt handlers are fast (as they are in -current but not in -stable
or in my version of -current).