Only poll the fifo for read events if the fifo is attached to a readable file descriptor.  Otherwise, the read end of a fifo might return that it is writable (which it isn't).
Only poll the fifo for write events if the fifo attached to a writable
file descriptor.  Otherwise, the write end of a fifo might return that
it is readable (which it isn't).

In the event that a file is FREAD|FWRITE (which is allowed by POSIX, but
has undefined behavior), we poll for both.

MFC after:	3 days