This is a massive cleanup of the fortune files, including removal of many duplicates, formatting/whitespace fixes, spelling fixes, and a host of others. Note: No fortunes were intentionally removed (other than duplicates) as part of this cleanup. In removing dupes I tried to be generous in not removing similarly worded fortunes that had even slightly different punch lines.
Duplicates were resolved as follows (in this order):
* + limerick		-> limerick
* + fortunes-o.real	-> fortunes-o.real
* + fortunes2-o		-> fortunes2-o
* + zippy		-> zippy
murphy + fortunes	-> fortunes

In correcting spelling I've tried to leave colloquial variations
(especially British'isms) intact, as well as to not "correct"
errors that are part of the joke.

In fortunes and fortunes2-o I combined a couple of fortunes that
were adjacent to each other that seemed like they should have
originally been combined, and split a couple that seemed like they
should have originally been split.

Miscellaneous:
1. Fixed underlining
2. Capitalized proper names
3. Removed quotation marks from "freestanding" quotes for consistency
4. Added quotation marks to QOTD fortunes that needed them
5. Corrected or added many attributions
6. Removed apostrophes from trailing 's' that are plural, not possessive
7. Updated the path to fortune stuff included in some of the jokes
8. Updated several fortunes to their original version, and added attribution
9. Split up compound words that ispell tripped over